+++ projects/testapp/include/comm_channels/udp_bridge.h
@@ -44,13 +44,13 @@ public:
     static void global_init();
     static void global_close();
 
-    void send(const uint8_t* buffer, int len, int flags = 0);
-    int receive(uint8_t* buffer, int len, int flags = 0);
+    void send(const uint8_t* buffer, size_t len, int flags = 0);
+    int receive(uint8_t* buffer, size_t len, int flags = 0);
     
     virtual void start();
     virtual void stop();
-    virtual void send(const uint8_t* buffer, int len) {send(buffer, len, 0);}
-    virtual int receive(uint8_t* buffer, int len) {return receive(buffer, len, 0);}
+    virtual void send(const uint8_t* buffer, size_t len) {send(buffer, len, 0);}
+    virtual int receive(uint8_t* buffer, size_t len) {return receive(buffer, len, 0);}
     
     void set_nonblocking();
     static void throw_system_error(const char* msg);
